Fix: remove events after destroy the fullscreen window.
GHOST_System::beginFullScreen will created a new window and put it in
the fullscreen mode. GHOST_System::endFullScreen will destroy it, but we
need to remove pending events related to the fullscreen window from the
event queue after we destroy the window. Otherwise, the system may
delivery events of an invalid window object to the application.